PROBLEM SOLVED!
It turns out that for some reason it simply doesn't run on my computer.
I tested the code with my txt files and it runs on multiple other
computers.  I will be spending the next while trying to figure out why
it doesnt work on my computer.  But at least I now know that my code has
been working for the past 3 days, it just never worked on my computer,
so I thought it was broken.