Everything is working
Ok earlier it wouldn't work on my computer.  Now the code is working on
my computer and other computers.  It turns out that the problem that I
had over the last three days was simply the fact that my txt files
weren't supposed to have the file extension on them. Will be adding
comments next